+++ tests/Feature/Console/DebugModules/Fixtures/GizmoModule/GizmoModuleFacade.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Console\DebugModules\Fixtures\GizmoModule;
+
+use Gacela\Framework\AbstractFacade;
+
+final class GizmoModuleFacade extends AbstractFacade
+{
+}