+++ tests/Feature/Console/DebugModules/Fixtures/WidgetModule/WidgetModuleConfig.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Console\DebugModules\Fixtures\WidgetModule;
+
+use Gacela\Framework\AbstractConfig;
+
+final class WidgetModuleConfig extends AbstractConfig
+{
+}