+++ tests/Feature/Console/DebugModules/Fixtures/WidgetModule/WidgetModuleFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Console\DebugModules\Fixtures\WidgetModule;
+
+use Gacela\Framework\AbstractFactory;
+
+final class WidgetModuleFactory extends AbstractFactory
+{
+}